Move SSE41 and AVX2 detection into Platform namespace
