+++ src/openrct2/drawing/LightFX.cpp
@@ -21,7 +21,6 @@
 #include "../ride/Ride.h"
 #include "../ride/RideData.h"
 #include "../ride/Vehicle.h"
-#include "../util/Util.h"
 #include "../world/Map.h"
 #include "Drawing.h"
