+++ src/openrct2/platform/Platform.h
@@ -124,6 +124,9 @@ namespace Platform
     uint32_t GetTicks();
 
     void Sleep(uint32_t ms);
+
+    bool SSE41Available();
+    bool AVX2Available();
 } // namespace Platform
 
 #ifdef __ANDROID__
